Refactored interface for persistence
